refactor(ui): move the render-cache invalidation chokepoint to diffview.go
round 2 left this as an open question rather than a finding.
invalidateRenderCaches clears diffRenderCache, but lived in annotate.go
alongside the annotation rows it also clears, while the cache, its type,
its clear method and every other use sat in diffview.go. A maintainer
adding a runtime color toggle found clear() with no visible caller and no
contract, and a missed invalidation is silent.

moved next to diffRenderCache. The three tests pinning it moved to
diffview_test.go with it, renamed off AnnotationRows since they cover
both memos, and each now asserts the diff blocks are dropped - none of
them did, which is how the function could have stopped clearing
renderCache without a test noticing.

the file-load test asserts no block from the previous file survives
rather than that the cache is empty: handleFileLoaded re-renders after
invalidating, so a warm cache for the new file is correct.
