+++ app/ui/diffview.go
@@ -933,6 +933,26 @@ func (c *diffRenderCache) put(idx int, flags lineRenderFlags, block string) {
 	c.filled[idx] = true
 }
 
+// invalidateRenderCaches clears both render memos: the annotation visual-row
+// slices and the per-line diff render blocks. callers: handleFileLoaded (per-file
+// annotation set changes), applyTheme (resolver colors change), cancelThemeSelect
+// (preview theme rebuilt the resolver), and handleBlameLoaded (blame data feeds
+// the gutter). Width and the other comparable render inputs self-invalidate via
+// the respective cache keys, so no call is needed on resize.
+//
+// This is the invalidation chokepoint for anything the caches read that a key cannot
+// capture. Four render inputs are not comparable and so cannot live in
+// globalRenderKey: the style resolver, file.blameData, file.highlighted and
+// file.intraRanges. A change to any of them is reflected only by calling this. The
+// last two are also covered today by loadSeq/fileName and the wordDiff key field
+// because of where they happen to be mutated, which is not something to rely on —
+// any new path that rebuilds the resolver, loads blame, re-highlights, or recomputes
+// intra-line ranges MUST call this.
+func (m *Model) invalidateRenderCaches() {
+	clear(m.annot.rowCache)
+	m.renderCache.clear()
+}
+
 // clear drops every cached block, keeping the allocated slices.
 func (c *diffRenderCache) clear() {
 	clear(c.blocks)
